Drawing Context Menu in the Middle
Drawing Context Menu at cursor position doesn't seem doable with the
current implementatoin of tview.Table. Also getting the relative
position of selection when the height is smaller than the row count
isn't feasible there are lot of bugs with that implementation as the
context menu is not drawn when the selection reaches the end of the list
or is out of the view. Instead a global context menu must be implemented
(this is what i think atleast for now) I think InteractiveView can go
back to it's previous implementation and context menu can be a global
thing.